refactor NotificationTip to TypeScript
